Update dag actions UI
