+++ airflow/ui/src/components/Clear/Run/ClearRunButton.tsx
@@ -17,7 +17,7 @@
  * under the License.
  */
 import { Box, useDisclosure } from "@chakra-ui/react";
-import { FiRefreshCw } from "react-icons/fi";
+import { LuRedo } from "react-icons/lu";
 
 import type { DAGRunResponse } from "openapi/requests/types.gen";
 import ActionButton from "src/components/ui/ActionButton";
@@ -36,7 +36,7 @@ const ClearRunButton = ({ dagRun, withText = true }: Props) => {
     <Box>
       <ActionButton
         actionName="Clear Dag Run"
-        icon={<FiRefreshCw />}
+        icon={<LuRedo />}
         onClick={onOpen}
         text="Clear Run"
         withText={withText}
