+++ airflow/ui/src/components/Clear/TaskInstance/ClearTaskInstanceDialog.tsx
@@ -18,7 +18,7 @@
  */
 import { Flex, Heading, VStack } from "@chakra-ui/react";
 import { useState } from "react";
-import { FiRefreshCw } from "react-icons/fi";
+import { LuRedo } from "react-icons/lu";
 
 import type { TaskInstanceResponse } from "openapi/requests/types.gen";
 import { ActionAccordion } from "src/components/ActionAccordion";
@@ -144,7 +144,7 @@ const ClearTaskInstanceDialog = ({ onClose, open, taskInstance }: Props) => {
                 }
               }}
             >
-              <FiRefreshCw /> Confirm
+              <LuRedo /> Confirm
             </Button>
           </Flex>
         </Dialog.Body>
