+++ airflow/ui/src/components/Graph/TaskNode.tsx
@@ -18,7 +18,7 @@
  */
 import { Box, Button, Flex, HStack, Text } from "@chakra-ui/react";
 import type { NodeProps, Node as NodeType } from "@xyflow/react";
-import { MdRefresh } from "react-icons/md";
+import { LuRedo } from "react-icons/lu";
 
 import { StateBadge } from "src/components/StateBadge";
 import TaskInstanceTooltip from "src/components/TaskInstanceTooltip";
@@ -92,7 +92,7 @@ export const TaskNode = ({
                   <StateBadge fontSize="xs" state={taskInstance.state}>
                     {taskInstance.state}
                   </StateBadge>
-                  {taskInstance.try_number > 1 ? <MdRefresh /> : undefined}
+                  {taskInstance.try_number > 1 ? <LuRedo /> : undefined}
                 </HStack>
               )}
             </Box>
