+++ airflow/ui/src/pages/Dag/Dag.tsx
@@ -17,9 +17,10 @@
  * under the License.
  */
 import { ReactFlowProvider } from "@xyflow/react";
-import { FiBarChart, FiCode, FiRotateCcw } from "react-icons/fi";
+import { FiBarChart, FiCode } from "react-icons/fi";
 import { LuChartColumn } from "react-icons/lu";
 import { MdDetails, MdOutlineEventNote } from "react-icons/md";
+import { RiArrowGoBackFill } from "react-icons/ri";
 import { useParams } from "react-router-dom";
 
 import { useDagServiceGetDagDetails, useDagsServiceRecentDagRuns } from "openapi/queries";
@@ -33,7 +34,7 @@ const tabs = [
   { icon: <LuChartColumn />, label: "Overview", value: "" },
   { icon: <FiBarChart />, label: "Runs", value: "runs" },
   { icon: <TaskIcon />, label: "Tasks", value: "tasks" },
-  { icon: <FiRotateCcw />, label: "Backfills", value: "backfills" },
+  { icon: <RiArrowGoBackFill />, label: "Backfills", value: "backfills" },
   { icon: <MdOutlineEventNote />, label: "Events", value: "events" },
   { icon: <FiCode />, label: "Code", value: "code" },
   { icon: <MdDetails />, label: "Details", value: "details" },
